Use size_t instead of int to store number of iconv conversion errors
The number represents a count of bytes, so size_t is the correct type
to use.

+++ src/cconv.c
@@ -15,7 +15,7 @@ struct cconv {
     size_t opos;
 
     size_t consumed;
-    int errors;
+    size_t errors;
 
     // Temporary input buffer
     char tbuf[16];
@@ -265,7 +265,7 @@ void cconv_flush(struct cconv *c)
     }
 }
 
-int cconv_nr_errors(const struct cconv *c)
+size_t cconv_nr_errors(const struct cconv *c)
 {
     return c->errors;
 }
